codeql: exclude vendored + sibling code by building it untraced; widen resize offsets
The paths-ignore in the previous commit did not work. It listed
third_party/** and the sibling checkouts, and the first scan still shipped
13 stb alerts to the Security tab. For a compiled language CodeQL analyses
what it observes the compiler compile — paths-ignore does not filter
compiled sources, which is documented and which I had wrong.

The mechanism that does work is to build that code before tracing starts:
anything already built when `init` runs is never extracted. So stb_image,
brotensor_core and brotensor_cpu are built in a step ahead of init, and
only broimage's own TUs are left for the traced build. bromath needs no
entry — header-only, no TUs of its own, and its headers should be analysed
where broimage includes them.

Worth excluding rather than dismissing: stb trips analysers by design, and
those findings are questions for upstream. A permanent wall of alerts
nobody can action is how a Security tab becomes something everyone ignores,
including on the day a real broimage bug lands in it. The brotensor targets
are analysed by brotensor's own workflow and would double-report here
against the wrong repo.

The resize kernel's offsets are widened to size_t. CodeQL flagged the two
weights[x * ksz] indices (they convert to size_type), but the same function
had four more int offsets feeding raw pointer arithmetic —
(y * src_w + x) * ch — which overflow at the same threshold and are worse:
UB in the pointer arithmetic rather than a merely wrong index. A 30k x 30k
RGB image reaches it. All six are widened together; the two that were
flagged were not the dangerous ones.

+++ src/geometric.cpp
@@ -180,17 +180,21 @@ void resize_axis_lanczos3(const float* src, int src_w, int src_h, int ch,
         }
     }
 
+    // The row terms are widened to size_t before multiplying. In int, an offset
+    // like (y * src_w + x) * ch overflows once the plane passes ~2^31 elements —
+    // a 30k x 30k RGB image gets there — and that is UB in the pointer
+    // arithmetic below, not merely a wrong index.
     if (axis_x) {
         for (int y = 0; y < src_h; ++y) {
             for (int x = 0; x < dst_w; ++x) {
-                const float* wrow = &weights[x * ksz];
+                const float* wrow = &weights[static_cast<size_t>(x) * ksz];
                 const int    lo   = first[x];
-                float* dp = dst + (y * dst_w + x) * ch;
+                float* dp = dst + (static_cast<size_t>(y) * dst_w + x) * ch;
                 for (int c = 0; c < ch; ++c) dp[c] = 0.0f;
                 for (int k = 0; k < ksz; ++k) {
                     const int sx = clampi(lo + k, 0, src_w - 1);
                     const float w = wrow[k];
-                    const float* sp = src + (y * src_w + sx) * ch;
+                    const float* sp = src + (static_cast<size_t>(y) * src_w + sx) * ch;
                     for (int c = 0; c < ch; ++c) dp[c] += sp[c] * w;
                 }
             }
@@ -198,15 +202,15 @@ void resize_axis_lanczos3(const float* src, int src_w, int src_h, int ch,
     } else {
         // dst layout: (dst_w aka new height, src_w, ch).
         for (int y = 0; y < dst_w; ++y) {
-            const float* wrow = &weights[y * ksz];
+            const float* wrow = &weights[static_cast<size_t>(y) * ksz];
             const int    lo   = first[y];
             for (int x = 0; x < src_w; ++x) {
-                float* dp = dst + (y * src_w + x) * ch;
+                float* dp = dst + (static_cast<size_t>(y) * src_w + x) * ch;
                 for (int c = 0; c < ch; ++c) dp[c] = 0.0f;
                 for (int k = 0; k < ksz; ++k) {
                     const int sy = clampi(lo + k, 0, src_h - 1);
                     const float w = wrow[k];
-                    const float* sp = src + (sy * src_w + x) * ch;
+                    const float* sp = src + (static_cast<size_t>(sy) * src_w + x) * ch;
                     for (int c = 0; c < ch; ++c) dp[c] += sp[c] * w;
                 }
             }
